requirements.txt
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
distro==1.9.0
Flask==3.1.1
flask-cors==6.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.10.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
openai==1.82.0
outcome==1.3.0.post0
packaging==25.0
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
pyscreeze>=0.1.28
PySocks==1.7.1
python-dotenv==1.1.0
pytweening==1.2.0
requests==2.32.3
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.30.0
trio-websocket==0.12.2
typing-inspection==0.4.1
typing_extensions==4.13.2
urllib3==2.4.0
webdriver-manager==4.0.2
websocket-client==1.8.0
Werkzeug==3.1.3
wsproto==1.2.0

# HTTP client & OpenAI
requests>=2.25
openai>=0.27

# browser automation
selenium>=4.8


# desktop automation + image matching

Pillow>=9.0         # PIL support for screenshots
opencv-python>=4.5  # enables confidence-threshold matching
